requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
redis
pydantic<2.0
pytest
httpx
